add LogVisualizerChart
